backend/requirements.txt
fastapi
uvicorn
sqlalchemy
garminconnect
garth
pydantic
python-dotenv
zhipuai
sniffio
